Fake data in separate files to reduce complexity
